add apk upgrade on container building
Sometimes HAProxy Ingress is released a few weeks after the upstream
haproxy container, or sometimes a new release is needed to fix a
security vulnerability. Adding `apk upgrade` during the container
building process guarantee that the most up to date dependencies, in the
date of the controller release, will be used in the generated image.

This should be merged to all maintained versions.