remove Path and Match from the HostPath struct
PathLink type is the way the controller creates an index of possible
incoming requests, and apply changes, filters and validations to the
request, as well as identify which backend the request should be
forwarded to.

Removing duplicated entities - in this case path and match from the
HostPath struct - is a first step to increase the ways a request can be
matched. Now the PathLink is the source of data used to classify
requests.