add a hash to pathlink
PathLink is used to distinguish two request match configuration. They
should be comparable and two distinct PathLinks with the same
configuration should match. Adding header host to the PathLink struct
made it not comparable anymore, because of that a hash is being added
to the struct, and the hash should be used to verify if two distinct
PathLinks match.