add support for service external name without port
A backend server is configured with the port number declared in the
service. However, services of type ExternalName doesn't need to have
a port declared, and since the v0.8 refactor, HAProxy Ingress refuses
to configure backends for services without port.

This update checks if the service is of type ExternalName and if the
ingress referencing this service uses a numeric port number. If true,
the port number declared in the ingress will be used as the port to
connect to the backend.