fix host redirects when acme is enabled
Acme validation happens on plain HTTP and should skip redirects, so that
the acme service (e.g. Let's Encrypt) can properly validate the domain
ownership. This wasn't properly happening with `redirect-from` and
`redirect-to` keys, which makes the signer to fail. This update skips
those redirects when acme is enabled and path matches the acme
challenge.