add warning if auth external svc isnt found
Auth External works by parsing the authentication service url and
looking for a pre-configured backend. The pre-configuration part works
by making a normalization in the k8s service port number, so different
inputs for the same service will output the same backend. This might
change the port name of the backend on named port configurations.

A proper fix is to make the backend creation accessible by the updater
code, so we don't need to pre-initialize the backend in the ingress
step, and don't need to guess the backend port name in the updater step.
So, for now, we'll improve logging and documentation.

This should be merged as far as v0.13, which implements the first
version of auth external.