fix path scoped annotation on service resources
Backend scoped annotations apply just once per service resource, and
have the most precedence. With this behavior in mind, service
annotations were being applied along with the annotation mapper
creation. This works nice with backend scoped annotations.

Path scoped, however, has the ability to apply distinct values to the
same backend. When path scoped annotations are applied via a service
resource, only the first path is being applied, along with annotation
mapper creation. This update ensures that the service annotations are
applied on every pathlink that points to it. This change will spread
service annotations over all referenced pathlinks, however this doesn't
end up with duplicated configurations because the config builder already
deduplicates configurations on the same key, with the same value, to the
same backend. Warning on annotation conflict isn't a problem as well, a
conflict is only warned on distinct values.