add source ip related config keys
Add options to declare a header name for the original X-Forwarded-For
header value, which currently is hard coded. Add also an option to
configure the header name which should receive the real source IP
despite of X-Forwarded-For configuration. This is actually hard coded
only on auth external configurations, now it's a global. Both options
can be removed by the cluster admin.

Other than that the X-Forwarded-For configuration were moved to the
frontend, since it's a global configuration, avoiding to replicate the
same config on every backend. It can be moved back to the backend in the
future if such configuration starts to be path scoped.