add haproxy types for frontend based auth external
External authentication services has the ability to inject headers on
the fly in the current request. Such header will be visible to the
backend server in the case the authentication succeed, and to the
caller/client in the case the authentication fails. This works that way
since v0.13.

Starting on v0.15 HAProxy Ingress has the ability to route requests
based on HTTP headers, along with the well known hostname+path supported
since the very first version. However headers added or updated by the
authentication service isn't visible to HAProxy Ingress routing rules,
this happens because the authentication service is called from the
backend declaration for performance reasons, but the routing rules must
be configured in the frontend due to its logic: it needs to choose a
backend based on hostname, path, and now HTTP headers.

This is a very first step on behalf of optionally bringing the
authentication service call to the frontend. This should start in a very
rudimentar way and with some performance impact if used on several
distinct pathlinks, but it's good enough on use cases that such
functionality is far more important than a small performance penalty.
Other than that: 1) this is optional, enabled by the user, and doesn't
make any difference on the former configuration; 2) future versions will
make improvements on such frontend configuration; 3) the performance
penalty grows based on the number of pathlinks, so adding a dozen or so
on moderately used proxy won't hurt, as much as adding one or two on a
highly used one.