add converter for the frontend based auth external
Add the converter implementation for the frontend based auth external.
The same backend code base was reused - the old one was moved to a
separated function without any change, and reused by frontend and
backend updaters. Note that we created a new class of annotation group,
duo, which made me wonder why we need such groups instead of a single
annotation key/value hash. Added a TODO in the code so we'll have a few
more time to think about it.