change owner of crt/key files to haproxy pid
HAProxy Ingress v0.14 and older starts as root by default and create
files as 0600 that should be read by HAProxy. When HAProxy runs as a
sidecar and stars as non root (default since its 2.4) it won't have
permission to read those files. This update makes files generated as
0600 (crt/key) having its owner to PID 99, which's the PID used by
HAProxy containers.