ensure predictable tcp by sorting endpoints
Dynamic update code relies on comparing tcp services structs in order to
verify whether any tcp service has changed or not, and if so, haproxy
need to be restarted. Endpoint list is the only tcp data that might lead
to differences in the case the same values are present, but in distinct
order. Because of that we're applying a sort in order to generate
predictable, hence comparable results.