redirect hosts only to domains with associated backends
Redirect-from configuration key defines the source hostnames that should
redirect to the one configured in the ingress manifest. This is not
properly working with ingress that adds more hostnames in the TLS
section because haproxy ingress adds all those hostnames as regular
ones. Adding as a regular hostname is needed in order to make haproxy
choose the right certificate in the TLS handshake.

The following change is making redirect-from apply redirects only on
hosts that have a backend association, which would be otherwise ending
up in a 404 error anyway.