move leader election to a self managed service
Controller-runtime's leader election works by killing the running
process when the leader is lost, consequently being restarted by the
runtime env. This works nice for a control plane component, but it'd
work badly with haproxy ingress because it currently embeds the data
plane as well on its deployment, making haproxy to restart on every
instance that looses an election. Controller-runtime behavior is not
configurable, so we need to implement our own leader election control.