add net bind capability to haproxy bin
HAProxy Ingress starts now as the non privileged user haproxy, and this
is the same user that starts the embedded haproxy process. Some
container runtimes do not configure the unprivileged_port_start param
from the default 1024 to 0, which makes the embedded haproxy process
fail to bind to 80/443 ports. Adding the net bind capability to the
haproxy binary allows it to bind to any tcp port, without running as a
privileged user, and without configuring unprivileged_port_start param.