add ssl-always-follow-redirect option
Controller v0.11 and older had an incorrect behavior compared with
ingress spec, regarding HTTPS configuration and SSL redirect.

v0.12 added ssl-always-add-https option that adds the host in the HTTPS
map when enabled, regardless of the tls attribute configuration. This
was a v0.11 and older behavior.

This update adds the ability to avoid SSL redirect when TLS was added
automatically. This was also a v0.11 and older behavior.