allows secure backend configuration from global
Some configuration keys were not designed to be made from global due to
the lack of a namespace name. Secure backend needs a namespace to read
secrets for a CA bundle and client certificate. This update allows to
provide a secret in the global ConfigMap by adding the namespace along
with the name of the secret.