fix memory leak on gateway reconciliation
Ingress and Gateway data are converted to haproxy frontend and backend
data via their converter implementations. Some configurations need data
from both sides, so frontend is linked to backend via a pathlink object.
A pathlink has path specific data, as well as the ID of the backend it
represents. A pointer from the pathlink to the ModeTCP field in the
backend, along with a backend data substitution used by shard mode, were
contributing to preserve old representations of a backend in memory,
depending on the type of the changes made to gateway resources.

This update breaks the link between a new resource and an old backend
representation, which fixes the memory leak.