check if haproxy reloaded already
master socket based reloads happen by issuing a reload command to
haproxy, which will fork a new process for the new configuration. This
process happens asynchronously, so there is an unexpected but possible
chance to talk with master via its socket before the reload starts.

HAProxy Ingress expects haproxy properly reflecting its internal model,
because of that it needs to wait for the reload to finish, otherwise
dynamic updates should fail due to outdated data.

This update adds a new check on this process: whenever a reload is going
to happen, the number of reloads is copied, and haproxy ingress will
continue to wait until it is incremented.