add context on socket calls
HAProxy Ingress shutdown is eventually racing during stop, leading to
the 25s controller manager timeout, taking more time for the controller
pod to be finalized. This is happening when sigint/sigterm happens just
before a new haproxy reload is issued: the signal closes the reconciler
context, which stops haproxy, making the reload check to wait forever
until haproxy is back again, leading to the timeout of the controller
manager.

This update changes the StopCh channel to a proper context, and uses
the context on all the socket calls.

Should be added to v0.15 branch.