create new event queues when leader is acquired (#1283)
Controller creates event queues for status and ACME updates, and they
run only on the leader. These event queues are not working anymore when
leader is lost and later acquired again.

This update recreates the event queue on every service behind leader
election. Logs and error handling are also being improved so similar
issues can be found easier in the future.

Should be added to v0.15