makes controller pod namespace and name mandatory (#1287)
A number of configurations require controller pod name and/or namespace
configured, starting to become harder to add and maintain all the
validations. We're configuring it mandatory now, since it is easy to
automate this configuration and default Helm chart already configures
it by default.

Need to add to the breaking backward compatibility list of v0.16.