read controller pod selector from owner (#1288)
Controller pod selector is used to list all pods that form the ingress
cluster, making it possible to find all the nodes where ingress
controller is running (used by default on ingress status) and peers.

Historically ingress controller uses pod labels to find all the other
controller pods, but this approach can lead to an inaccurate result in
the case some of the labels uniquely identify the pod.

This update changes the approach to use the owner selector instead,
being either a DaemonSet or a ReplicaSet. The owner will always have an
accurate selector to find all their pods.

should be added to v0.15.