configure more than one http/s ports
Adds the ability to configure the HTTP/s frontends with more than one
HTTP port, allowing to multiplex distinct groups of hostnames without
colliding each other. This is a pre-requisite to improve Gateway API
support.

Several places of HAProxy Ingress code uses the dual and plain HTTP/s
model as a premise, so this is a non trivial refactor. Starting the
refactor from the top of the internal HAProxy model representation.

TODO list includes:

* [x] Move Hosts and Frontend types under a new Frontends type
* [x] Improve PathLink type to support distinct Frontends
* [ ] Move Bind configurations specific to frontends under the Frontend type
* [ ] Split the single HTTP/s frontend representation into two separated HTTP and HTTPs entries
* [ ] Support a list of HTTP/s frontends
* [ ] Add frontends support into Ingress API
* [ ] Add listener.Port support into Gateway API