fix reading backend ca certificate from file
TLS backend can be configured by adding a CA secret (verify server's
certificate) and/or a Crt secret (server verify client's certificate).
A former controller version was reading the namespace of a secret from
the ingress that configures the annotation, leading to nil dereference
when the value is configured globally. This was fixed by adding a new
step, that verifies if the config is global or an annotation, and if
the provided name has already a name or just the name of the secret.

This new step however broke the proto://content approach added on the
get secret code because it happens before getting the secret, and the
proto://content syntax is incompatible with the ns/name one. This
failed deployments that read ca or crt from the file system.

Should be merged as far as v0.13.