fix xfp header on fronting proxy shared port
Fronting proxy configuration is incorrectly handling the
X-Forwarded-Proto header sent to the backend if fronting proxy is
configured in the same HTTP port: the frontend configuration overwrites
the header value before its value is interpreted by the backend.

This update configures a txn var in the frontend before update the
header value, so the value provided by the user is correctly handled.

Should be merged as far as the v0.13 branch.