move gateway api version check to the cache
Simplifies the Cache facade by moving the Gateway version check from
the converter to the cache. Instead of making the converter to identify
which cache method should be called, the cache method itself checks the
version and converts the api version if needed.

Gateway converter implementation has also a huge improvement: since the
code needs to deal with a single version per API, code could be
refactored in a way it has a better reuse.

An internal behavior was changed as well: whenever an undeployed/
unsupported api version endpoint is used, getlist() will return an
empty list and get() will return a 404 compatible error, so converters
just need to make their ordinary calls and handle the same errors as
usual without the need to check supported api versions.