fix full-sync trigger
Full sync is used whenever controller needs to parse all the Ingress and
Gateway API resources, e.g. when a global is changed. Some internal
events like acquire leader also leads to a full sync, but in this case
there is no changed object.

Our previous approach to run full-sync without object being changed was
parameterize a full-sync flag in the request object, but this might
misbehave in the case there is also a non full-sync event in the queue,
and converter needs to deal with the changed objects on the full-sync
event.

So instead of hacking the internal work queue from the
controller-runtime's controller, we're now using the generic event that
can be triggered via a channel.

This fix should be merged up to v0.15.