make status update synchronous (#1330)
Changing status update from the previous async approach to a synchronous
one. Although a bit faster, running it async has the disadvantage of the
need to deal with conflicts.

Running it synchronously, on the other hand, does not have a significant
performance impact: changes are the exception during the event loop, so
they should happen fast. Batch updates like address changes happen on a
dedicated work queue, so they do not block or delay other processing.