fix status update during shutdown
Address status are removed from ingress resources in the case controller
detects it is the last instance running. This behavior is configurable.
The status update during the shutdown is not working because:

* Controller is not the leader anymore, and status updates only happen
  in the leader;
* Internal context is already cancelled, making the status not to be
  updated.

This change is parameterizing status update so it ignores the leader
state during the shutdown event. A new context is also being used.

This fix should be merged up to v0.15.