add http passthrough config keys (#1340)
Introducing http passthrough configuration, better aligned with multi
frontends feature. Ports are now configured globally only. A flag is
introduced in the frontend scope, so custom frontends can be configured
as http passthrough from an ingress annotation.

This is also the replacement for fronting-proxy-port. Legacy
fronting-proxy-port configuration should continue to work, except on
the unsafe use case where a mix of regular and a passthrough
configuration shares the same TCP port.