add option to always create default frontends (#1342)
* add option to always create default frontends

Up to v0.16 there was a single pair of HTTP and HTTPS frontends, it was
always created despite of being used. Since v0.17 there is the concept
of list of frontends, and since then the default frontends are part of
this list, they are created only if used.

This new option adds the ability to always create the default frontends,
just like it happened on older versions, so admins can opt to bring back
this behavior on clusters that eventually have no resources but need to
always bind HTTP and HTTPS ports.

* remove empty frontends on partial parsing

* fix frontend ports naming

* add integration test