add connection timeout command-line option
Add connection timeout option, which parameterize how much time to wait
for HAProxy to respond for an admin or master command. This is specially
useful on HAProxy v2.7+, whose reload command is synchronous, and should
go beyond the hardcoded 5s.

Should be merged up to v0.14, which fixes this behavior if using a newer
HAProxy version as a sidecar.