add custom http frontend config keys (#1341)
http-frontend(s) config keys adds the concept of frontend ID, which is
now a mandatory configuration whenever an ingress resource tries to
configure a frontend scoped key, like http passthrough or proxy protocol
using ingress annotations. This is done that way so an ingress does not
have the permission to make global changes on the default http and https
frontends.

Adding also `allow-local-bind` so listening bind customizations are
denied by default, even if ingress resources have the ability to
configure and customize frontends.