check if process is the same waiting for reload
HAProxy up to 2.6 runs the reload command async, so we need to wait
until we identify that haproxy finished reloading, succeeding or not.
This is done by checking if the number of reloads increased. We however
does not check if the process continues to be the same, in that case
we'd see the reload count decreasing. This update checks not only the
number of reloads, but also the process ID of the previous and the
current HAProxy instance.